requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
sqlparse==0.5.1
tzdata==2024.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-passwordreset==1.3.0
djangorestframework-simplejwt==5.3.1
drf-nested-forms==1.1.8
email-validator==2.1.0.post1
numpy==1.26.1
idna==3.7
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
urllib3==2.2.1